requirements.txt
PyQt5>=5.15.0
PyQtChart>=5.15.0
pandas>=1.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
gspread>=5.0.0
google-auth>=2.0.0
numpy>=1.20.0
openai>=1.0.0
python-dotenv>=0.19.0
tiktoken>=0.3.0
pymongo>=4.0.0